fix(content-mapper): harden Vue semantic links
