+++ crates/vize_canon/src/batch/virtual_project/content_mapper_tests.rs
@@ -73,6 +73,35 @@ const café = ref(1)
 }
 
 #[test]
+fn keeps_tsx_semantic_link_ranges_aligned_after_jsx_reference_prefix() {
+    let source = r#"<script setup lang="tsx">
+import { ref } from 'vue'
+const café = ref(<span />)
+</script>
+<template>{{ café }}</template>
+"#;
+    let result = generate_vue_content_mapper_transform(Path::new("UnicodeRef.vue"), source)
+        .expect("transform");
+    let source_starts = result
+        .semantic_links
+        .iter()
+        .filter(|link| {
+            &result.text.as_str()[link.source_start..link.source_start + link.source_length]
+                == "café"
+                && &result.text.as_str()[link.target_start..link.target_start + link.target_length]
+                    == "café"
+        })
+        .count();
+
+    assert!(
+        source_starts > 0,
+        "expected both semantic-link endpoints to point at generated café ranges:\ntext:\n{}\nlinks:\n{:#?}",
+        result.text,
+        result.semantic_links
+    );
+}
+
+#[test]
 fn maps_synthetic_prop_bindings_to_the_authored_declaration() {
     let source = r#"<script setup lang="ts">
 defineProps<{ count: number }>();
